Issue #6 Netbox backend prefix without gateway tag
When filtering objects in the Netbox API using a tag that does not
exist, the API returns a 400.  This is different to the situation
where the tag does exist, but the full filter list matches 0 results
which results in a 200 and an empty result set.

This change handles the exception raised by the 400.